requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
cloudinary==1.15.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
coveralls==1.5.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.5
django-cors-middleware==1.2.0
django-extensions==2.1.5
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.9.1
docopt==0.6.2
entrypoints==0.3
flake8==3.7.5
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mock==2.0.0
oauthlib==3.0.1
openapi-codec==1.3.2
pbr==5.1.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pyflakes==2.1.0
PyJWT==1.4.2
pylint==2.2.2
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.10.0
social-auth-app-django==3.1.0
social-auth-core==3.0.0
tox==3.7.0
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.3.0
whitenoise==4.1.2
wrapt==1.11.1
